+++ src/Application/DependencyInjection.cs
@@ -7,12 +7,14 @@ public static class DependencyInjection
 {
     public static IServiceCollection AddApplication(this IServiceCollection services)
     {
-        services.AddSingleton<ChatCompletionsService>();
         services.AddScoped<EmbeddingNeighboursService>();
-        services.AddSingleton<PruningService>();
-        services.AddSingleton<TokenService>();
         services.AddScoped<MessageHandler>();
         services.AddScoped<RelevantRulesService>();
+        services.AddScoped<LeaderboardService>();
+        
+        services.AddSingleton<ChatCompletionsService>();
+        services.AddSingleton<PruningService>();
+        services.AddSingleton<TokenService>();
         services.AddSingleton<OpenAiServiceFactory>();
 
         return services;
