+++ src/Domain/Entities/LeaderboardModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Domain;
+﻿namespace Domain.Entities;
 
 public class LeaderboardModel
 {
