+++ src/WebUI/Models/OrderedLeaderboardUser.cs
@@ -1,0 +1,6 @@
+﻿namespace WebUI.Models;
+
+public class OrderedLeaderboardUser : LeaderboardUser
+{
+    public int Position { get; set; }
+}
